requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
folium>=0.15.0
streamlit-folium>=0.18.0
plotly>=5.18.0
google-genai>=0.3.0
python-dotenv>=1.0.0
